allow larger values

+++ schema.sql
@@ -6,6 +6,6 @@ create table data (
     data_retrieval_method varchar(1000),
     metric varchar(200),
     units varchar(100),
-    value float(28,14),
+    value float(35,14),
     notes varchar(2000) DEFAULT NULL
 ) ENGINE=InnoDB AUTO_INCREMENT=15239276 DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
